pyproject.toml
[project]
name = "grader-dumper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "lxml==5.3.0",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
]
